Move Userdev component into VersionFormattedCode and document in CONTRIBUTING.md
