+++ docs/paper/dev/getting-started/userdev.mdx
@@ -45,13 +45,13 @@ Portal](https://plugins.gradle.org/plugin/io.papermc.paperweight.userdev).
 
 :::
 
-<UserdevVersionCode language={"kotlin"} title={"build.gradle.kts"}>
+<VersionFormattedCode language={"kotlin"} title={"build.gradle.kts"}>
 ```
 plugins {
     id("io.papermc.paperweight.userdev") version "%%_USERDEV_VER_%%"
 }
 ```
-</UserdevVersionCode>
+</VersionFormattedCode>
 
 The latest version of `paperweight-userdev` supports dev bundles for Minecraft 1.17.1 and newer, so it's best practice to keep it up to date!
 Only the latest version of `paperweight-userdev` is officially supported, and we will ask you to update first if you are having problems with old versions.
